training/requirements.txt
opencv-python
torch
torchvision
tensorboardX
yacs
albumentations
sam-pytorch
tqdm
tabulate
onnx
onnxruntime
onnxsim
